+++ qiniu/region.py
@@ -23,13 +23,15 @@ class Region(object):
             home_dir=None,
             scheme="http",
             rs_host=None,
-            rsf_host=None):
+            rsf_host=None,
+            api_host=None):
         """初始化Zone类"""
         self.up_host = up_host
         self.up_host_backup = up_host_backup
         self.io_host = io_host
         self.rs_host = rs_host
         self.rsf_host = rsf_host
+        self.api_host = api_host
         self.home_dir = home_dir
         self.host_cache = host_cache
         self.scheme = scheme
@@ -91,6 +93,20 @@ class Region(object):
         rsf_hosts = bucket_hosts['rsfHosts']
         return rsf_hosts[0]
 
+    def get_api_host(self, ak, bucket, home_dir=None):
+        from .config import get_default, is_customized_default
+        if self.api_host:
+            return self.api_host
+        if is_customized_default('default_api_host'):
+            return get_default('default_api_host')
+        if home_dir is None:
+            home_dir = os.getcwd()
+        bucket_hosts = self.get_bucket_hosts(ak, bucket, home_dir)
+        if 'apiHosts' not in bucket_hosts:
+            bucket_hosts = self.get_bucket_hosts(ak, bucket, home_dir, force=True)
+        api_hosts = bucket_hosts['apiHosts']
+        return api_hosts[0]
+
     def get_up_host(self, ak, bucket, home_dir):
         bucket_hosts = self.get_bucket_hosts(ak, bucket, home_dir)
         if 'upHosts' not in bucket_hosts:
@@ -123,39 +139,38 @@ class Region(object):
         if not force and len(bucket_hosts) > 0:
             return bucket_hosts
 
-        hosts = {}
-        hosts.update({self.scheme: {}})
+        hosts = compat.json.loads(self.bucket_hosts(ak, bucket)).get('hosts', [])
 
-        hosts[self.scheme].update({'up': []})
-        hosts[self.scheme].update({'io': []})
+        if type(hosts) is not list or len(hosts) == 0:
+            raise KeyError("Please check your BUCKET_NAME! Server hosts not correct! The hosts is %s" % hosts)
 
-        if self.up_host is not None:
-            hosts[self.scheme]['up'].append(self.scheme + "://" + self.up_host)
+        region = hosts[0]
 
-        if self.up_host_backup is not None:
-            hosts[self.scheme]['up'].append(
-                self.scheme + "://" + self.up_host_backup)
+        default_ttl = 24 * 3600  # 1 day
+        region['ttl'] = region.get('ttl', default_ttl)
 
-        if self.io_host is not None:
-            hosts[self.scheme]['io'].append(self.scheme + "://" + self.io_host)
-
-        if len(hosts[self.scheme]) == 0 or self.io_host is None:
-            hosts = compat.json.loads(self.bucket_hosts(ak, bucket))
-        else:
-            # 1 year
-            hosts['ttl'] = int(time.time()) + 31536000
-        try:
-            scheme_hosts = hosts[self.scheme]
-        except KeyError:
-            raise KeyError(
-                "Please check your BUCKET_NAME! The UpHosts is %s" %
-                hosts)
         bucket_hosts = {
-            'upHosts': scheme_hosts['up'],
-            'ioHosts': scheme_hosts['io'],
-            'rsHosts': scheme_hosts['rs'],
-            'rsfHosts': scheme_hosts['rsf'],
-            'deadline': int(time.time()) + hosts['ttl']
+            'upHosts': [
+                '{0}://{1}'.format(self.scheme, domain)
+                for domain in region.get('up', {}).get('domains', [])
+            ],
+            'ioHosts': [
+                '{0}://{1}'.format(self.scheme, domain)
+                for domain in region.get('io', {}).get('domains', [])
+            ],
+            'rsHosts': [
+                '{0}://{1}'.format(self.scheme, domain)
+                for domain in region.get('rs', {}).get('domains', [])
+            ],
+            'rsfHosts': [
+                '{0}://{1}'.format(self.scheme, domain)
+                for domain in region.get('rsf', {}).get('domains', [])
+            ],
+            'apiHosts': [
+                '{0}://{1}'.format(self.scheme, domain)
+                for domain in region.get('api', {}).get('domains', [])
+            ],
+            'deadline': int(time.time()) + region['ttl']
         }
         home_dir = ""
         self.set_bucket_hosts_to_cache(key, bucket_hosts, home_dir)
@@ -165,9 +180,6 @@ class Region(object):
         ret = {}
         if len(self.host_cache) == 0:
             self.host_cache_from_file(home_dir)
-
-        if self.host_cache == {}:
-            return ret
 
         if key not in self.host_cache:
             return ret
@@ -211,7 +223,7 @@ class Region(object):
         uc_host = UC_HOST
         if is_customized_default('default_uc_host'):
             uc_host = get_default('default_uc_host')
-        url = "{0}/v1/query?ak={1}&bucket={2}".format(uc_host, ak, bucket)
+        url = "{0}/v4/query?ak={1}&bucket={2}".format(uc_host, ak, bucket)
         ret = requests.get(url)
         data = compat.json.dumps(ret.json(), separators=(',', ':'))
         return data